Bosses Update
Both bosses are in!
Just have to add in yiohan's level, put a fast travel from the end of Bulwark's arena to the start of yiohan's level, and put Regulus in the end of yiohan's level, and also add a whole bunch of instruction UI along the way/ phase shift and time lapse unlocks, then we should be done for the most part!